more small changes

+++ src/schedulers/dynamic/scheduler.c
@@ -338,7 +338,7 @@ void Scheduler_run(Scheduler *untyped_self) {
       going_to_shutdown = true;
     }
 
-    for (int i = 0; i < self->env->net_bundles_size; i++) {
+    for (int i = 0; i < (int)self->env->net_bundles_size; i++) {
       if (self->env->net_bundles[i]->net_channel->mode == NETWORK_CHANNEL_MODE_POLLED) {
         ((PolledNetworkChannel *)self->env->net_bundles[i]->net_channel)
             ->poll((PolledNetworkChannel *)self->env->net_bundles[i]->net_channel);
